# Variable length types scan speedup