I'm pulling in a rework of the variable length types (VLT for short) scan functions, salient points:

* Implement basic incoming memory length-checking, which offers a bit more safety and fixes the VLT returning from deletion of #204 .
* Massively optimize both small lengths (time -60%) and immediately not matching big lengths (-50%)'s scan time
* Rework bytearray wildcards, for another 15% speedup stacked on
* Finally closes #226 , as now VLT length is fully exploited

This PR features a fair share of type punning via pointer casting, I'm 99.9% sure that I've not introduced unaligned access on ARM, because the pointers I recast and dereference have enough alignment anyway, but I'd appreciate if someone could check on ARM.