@12345ieee 2017-08-15
Did somebody have a look? 
I tried an ARM VM, but couldn't get any program to crash on unaligned access, so I can't tell if it would actully work.