@sriemer 2017-08-16
Testing on physical HW would be best. A Raspberry Pi is ideal for that. The code looks good but I'd like to check on my Raspberry Pi 2.