@12345ieee 2017-08-16
If you can, I'd really appreciate that, @sriemer .

The speed improvements should be evident in any architecture, assuming nothing crashes.