@sriemer 2017-08-21
On the Raspberry Pi I see only small performance improvement. Nothing crashes but I also see a regression for strings. While a 7 char string is at 0.3s while scanning in a bash process, an 8 char string is at 1.7s. With the old code, that only happens from 9 chars onward. That needs to be fixed.