@12345ieee 2017-08-21
Thanks for testing.

To be clear:

1-7: new is faster?
8: new is slower
9- : new is faster?