@12345ieee 2017-08-21
Also, you're testing non-matching values, right? Otherwise all the speed gains are muddled in the much bigger time needed to add the match to the swaths.