@12345ieee 2017-08-21
I just tested on my laptop non-matching strings of length 7,8,9.
In all 3 cases, sys time is the same and user time is roughly half with the new code.

Can you ensure that your kernel actually traps on misalignment, instead of executing with a speed penalty, via `/proc/cpu/alignment`?