@sriemer 2017-08-21
Yes, of cause I'm testing non-matching stuff. Yep, also can't reproduce on x86_64.

Okay, I had "2 (fixup)" set before. I've set it to 4 and the cases which were slow now fail with SIGBUS. That means starting at 8 chars with this PR and starting at 9 chars with master.

Before:
```
User:           11239424
System:         0 (  (null))
Skipped:        0   
Half:           0   
Word:           0   
DWord:          11239424
Multi:          0   
User faults:    2 (fixup)
```
After:
```
User:           11239439
System:         0 (  (null))
Skipped:        0   
Half:           0   
Word:           0   
DWord:          11239424
Multi:          0   
User faults:    4 (signal)
```