@12345ieee 2017-08-21
Ok, so it's clearly the cast to `uint64_t`, which happens in my PR starting at 8, but on master starting at 9.

It's clearly fault of the uservalue string, which is not aligned.

Can you test with bytearrays? I allocate them with `malloc()`, so they should be always aligned at 16 bytes boundary.

I'm also tagging this as "bug", as the current code fails on strings over 8.