@sriemer 2017-08-21
The bytearrays work fine no matter which size.