@12345ieee 2017-08-21
Great to know.

I'll prepare an extra fix to allocate strings the same way bytearrays are allocated.

Can you actually see the speed difference for bytearrays? Or even on amd64? Here the difference is big and I don't think it's an artifact of my setup.