@sriemer 2017-08-21
Yes, the performance difference is big. I've now compiled my memfake tool with 200 MiB of zeroed memory to get a good measurement on my Raspberry Pi 2.

**bytearrays**
old: always 16.9s

new:
1..4 bytes: 10.7s            (37% less)
5..8 bytes: 12.5s
9..16 bytes: 14.8s
more bytes: time is gradually increasing until 16.9s where values are the same as with old scanmem

**strings**
old: 1..8 chars: 15.3s
new: 1..7 chars: 10.8s      (30% less)