@12345ieee 2017-08-21
I've force-pushed said fix: 32637da

String scan speed got a ~5% faster on amd64 as well and now matches the bytearray time.

Please have a look and test it.