@sriemer 2017-08-24
Hmm, it's a pity that Coverity doesn't find such stuff. With the new code everything is fine now. /proc/cpu/alignment doesn't complain about scanmem any more.

This can be merged now. Thanks!