@sriemer 2017-08-24
I like the new "rebase and merge". 👍