@12345ieee 2017-08-24
Yeah, it was about time github got a decent merge option from the site.

The only issue is that it adds you as committer and uses your default github email, which is not the same one I use for this project, so I keep merging manually.