@12345ieee 2017-08-26
@sriemer , I've seen you didn't commit the string bugfix in the 0.16 branch.
Have you changed idea about maintaining 0.16 as a bugfix branch or did you simply forget?