@sriemer 2017-08-28
Right, the 0.16 branch is the bugfix branch and should be maintained as that. I just didn't have enough time for it that day. Feel free to commit yourself.