@sriemer 2017-08-29
Thanks, LGTM. I've committed 9534cdb8f89820e047c937fdde21b3aff41b6bec to show that this is not v0.16.1 any more but preparation for 0.16.2 instead.