@12345ieee 2017-08-29
On the release topic, I'd like to accelerate a bit towards 0.17 as well.
Could you take a look at #262 ? #186 is the biggest roadblock towards 0.17 and I'd like some feedback on what we really want in `scanmem.h`, #262 should go in the right direction.